# backend/docs/swagger.yaml
components:
  schemas:
...
    dto.CreateGrade:
      properties:
        comment:
          maxLength: 255
          type: string
        curriculum_id:
          type: string
        student_id:
          type: string
        value:
          type: integer
      required:
      - curriculum_id
      - student_id
      - value
      type: object
    dto.CreateGroup:
      properties:
        admission_year:
          maximum: 2100
          minimum: 1900
          type: integer
        name:
          maxLength: 255
          minLength: 1
          type: string
        specialty_id:
          type: string
        study_form:
          $ref: '#/components/schemas/repositories.StudyForm'
      required:
      - admission_year
      - name
      - specialty_id
      - study_form
      type: object
...
    dto.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
    dto.Grade:
      properties:
        author_id:
          type: string
        comment:
          type: string
        created_at:
          type: string
        curriculum_id:
          type: string
        id:
          type: string
        student_id:
          type: string
        updated_at:
          type: string
        value:
          type: integer
      type: object
    dto.Group:
      properties:
        admission_year:
          type: integer
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        specialty_id:
          type: string
        study_form:
          $ref: '#/components/schemas/repositories.StudyForm'
      type: object
...
    dto.UpdateGrade:
      properties:
        comment:
          maxLength: 255
          type: string
        curriculum_id:
          type: string
        student_id:
          type: string
        value:
          type: integer
      type: object
    dto.UpdateGroup:
      properties:
        admission_year:
          maximum: 2100
          minimum: 1900
          type: integer
        name:
          maxLength: 255
          minLength: 1
          type: string
        specialty_id:
          type: string
        study_form:
          $ref: '#/components/schemas/repositories.StudyForm'
      type: object
...
    repositories.StudyForm:
      enum:
      - full_time
      - evening
      - extramural
      - full_time
      - evening
      - extramural
      - full_time
      - evening
      - extramural
      type: string
      x-enum-varnames:
      - StudyFormFullTime
      - StudyFormEvening
      - StudyFormExtramural
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: REST API образовательной системы Edutrack.
  title: Edutrack API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /grades:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.Grade'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить список оценок
      tags:
      - grades
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateGrade'
                description: Данные оценки
                summary: request
        description: Данные оценки
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Grade'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать оценку
      tags:
      - grades
  /grades/{id}:
    delete:
      parameters:
      - description: Идентификатор оценки
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить оценку
      tags:
      - grades
    get:
      parameters:
      - description: Идентификатор оценки
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Grade'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить оценку по идентификатору
      tags:
      - grades
    patch:
      parameters:
      - description: Идентификатор оценки
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateGrade'
                description: Данные оценки
                summary: request
        description: Данные оценки
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Grade'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Частично обновить оценку
      tags:
      - grades
  /groups:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.Group'
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить список групп
      tags:
      - groups
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateGroup'
                description: Данные группы
                summary: request
        description: Данные группы
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Group'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Создать группу
      tags:
      - groups
  /groups/{id}:
    delete:
      parameters:
      - description: Идентификатор группы
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Удалить группу
      tags:
      - groups
    get:
      parameters:
      - description: Идентификатор группы
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Group'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Получить группу по идентификатору
      tags:
      - groups
    patch:
      parameters:
      - description: Идентификатор группы
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateGroup'
                description: Данные группы
                summary: request
        description: Данные группы
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Group'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.Error'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Частично обновить группу
      tags:
      - groups
...
servers:
- url: /api